+++ packages/store/src/execution/internal-ngxs-execution-strategy.ts
@@ -1,12 +1,10 @@
-import { Injectable, Inject } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 
 import { NgxsExecutionStrategy, NGXS_EXECUTION_STRATEGY } from './symbols';
 
 @Injectable({ providedIn: 'root' })
 export class InternalNgxsExecutionStrategy implements NgxsExecutionStrategy {
-  constructor(
-    @Inject(NGXS_EXECUTION_STRATEGY) private _executionStrategy: NgxsExecutionStrategy
-  ) {}
+  private _executionStrategy = inject(NGXS_EXECUTION_STRATEGY);
 
   enter<T>(func: () => T): T {
     return this._executionStrategy.enter(func);