+++ packages/router-plugin/internals/src/symbols.ts
@@ -2,8 +2,6 @@ import { InjectionToken } from '@angular/core';
 
 declare const ngDevMode: boolean;
 
-const NG_DEV_MODE = typeof ngDevMode !== 'undefined' && ngDevMode;
-
 export const enum NavigationActionTiming {
   PreActivation = 1,
   PostActivation = 2
@@ -14,12 +12,12 @@ export interface NgxsRouterPluginOptions {
 }
 
 export const ɵUSER_OPTIONS = new InjectionToken<NgxsRouterPluginOptions | undefined>(
-  NG_DEV_MODE ? 'USER_OPTIONS' : '',
+  typeof ngDevMode !== 'undefined' && ngDevMode ? 'USER_OPTIONS' : '',
   { providedIn: 'root', factory: () => undefined }
 );
 
 export const ɵNGXS_ROUTER_PLUGIN_OPTIONS = new InjectionToken<NgxsRouterPluginOptions>(
-  NG_DEV_MODE ? 'NGXS_ROUTER_PLUGIN_OPTIONS' : '',
+  typeof ngDevMode !== 'undefined' && ngDevMode ? 'NGXS_ROUTER_PLUGIN_OPTIONS' : '',
   { providedIn: 'root', factory: () => ({}) }
 );